frontend/pyproject.toml
[project]
name = "step1x3d-frontend"
version = "1.0.0"
dependencies = [
    "streamlit>=1.40.0",
    "requests>=2.32.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "trimesh>=4.0.0",
    "pygltflib>=1.16.0",
]

[tool.setuptools]
py-modules = []

